+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer/file/EmojiMapping.java
@@ -1,6 +1,5 @@
 package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.file;
 
-import java.io.IOException;
 import java.nio.file.Path;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
@@ -13,12 +12,6 @@ public interface EmojiMapping {
 	 */
 	@Nullable
 	String getFileEmoji(Path file);
-
-	/**
-	 * Get the emoji to display for the given file, in case of error.
-	 */
-	@Nullable
-	String getErrorEmoji(Path file, IOException exc);
 
 	static EmojiMapping createDefault() {
 		return DefaultEmojiMapping.builder().build();